+++ CHANGELOG.md
@@ -16,6 +16,7 @@ SPDX-License-Identifier: GPL-3.0-or-later
 
 ### Added
 
+- relay/api: add resource `GET /api/scripts`
 - relay: add option relay.network.unix_socket_permissions ([#2317](https://github.com/weechat/weechat/issues/2317))
 - script: add info "script_languages"